Skeleton of molecules, clean warning, move imports
