+++ src/pages/Test.jsx
@@ -4,7 +4,6 @@ import styled from '@emotion/styled';
 import LupisImage from '../image/Lupis.png';
 import Tipography from '../molecules/Tipography';
 import Buttons from '../molecules/Buttons';
-import AppRouter from '../routers/AppRouter';
 import {Header} from '../molecules/Header';
 import {SearchBar} from '../molecules/SearchBar';
 import {OrdenItem} from '../molecules/OrdenItem';
@@ -17,48 +16,26 @@ import {TotalSalesDay} from '../molecules/TotalSalesDay';
 import TextAboutUs from '../molecules/TextAboutUs';
 import HomeImage from '../image/homeImage.svg';
 import {ItemsCategoryPurchaseReceipt} from '../molecules/ItemsCategoryPurchaseReceipt';
-import HeadearPurchaseReceipt from '../molecules/HeadearPurchaseReceipt';
 import {ProductCard} from '../molecules/ProductCard';
-import {TotalPurchaseReceipt} from '../molecules/TotalPurchaseReceipt';
-import {FooterPurchaseReceipt} from '../molecules/FooterPurchaseReceipt';
-import {Itemspurchasereceipt} from '../molecules/ItemsPurchaseReceipt';
 import {NotificationModal} from '../molecules/NotificationModal';
 import DescriptionPrice from '../atoms/DescriptionPrice';
 
 function Test() {
-  const onClickSearch = (searchText) => {
-    console.log(searchText);
-  };
+  // const onClickSearch = (searchText) => {
+  // console.log(searchText);
+  // };
 
   const handleDeleteItem = () => {
-    console.log('Delete item');
+    // console.log('Delete item');
   };
 
   const [orders] = useState({name: 'Órdenes', cant: 45});
   const [raised] = useState({name: 'Recaudado', cant: '$2800'});
   const [products] = useState({name: 'Productos', cant: 57});
-  const headerPurchase = {
-    date: '01/01/2020',
-    hour: '10:19 P.M.',
-    order: 0,
-  };
 
   const Image = styled.img`
     width: 50rem;
   `;
-
-  const purchase = {
-    paymentMethod: 'Tarjeta Terminada ***5678',
-    total: 225.0,
-  };
-
-  const itemPurchase = {
-    category: 'Bebidas',
-    products: [
-      {id: 1, name: 'Café Mocka', cant: 2, total: 110},
-      {id: 2, name: 'Café ', cant: 4, total: 220},
-    ],
-  };
 
   const productData = {
     image:
@@ -76,11 +53,6 @@ function Test() {
     <>
       <ProductCard image={productData.image} price={productData.price} title={productData.title} />
 
-      <HeadearPurchaseReceipt
-        paymentDate={headerPurchase.date}
-        paymentHour={headerPurchase.hour}
-        paymentNumber={headerPurchase.order}
-      />
       <Header />
 
       <section>
@@ -93,13 +65,12 @@ function Test() {
       </section>
 
       <TextAboutUs />
-      <Itemspurchasereceipt />
+
       <ItemsCategoryPurchaseReceipt
         category={itemPurchase.category}
         products={itemPurchase.products}
       />
-      <TotalPurchaseReceipt paymentMethod={purchase.paymentMethod} total={purchase.total} />
-      <FooterPurchaseReceipt />
+
       <Image src={HomeImage} />
       <TotalSalesDay orders={orders} products={products} raised={raised} />
       <CardPayment />
